Remove Backend trait (#6407)
